requirements.txt
kafka-python==2.0.2
pydantic>=2.9.0
optimum[openvino,nncf]
transformers
torch
scipy